+++ CHANGELOG.md
@@ -91,6 +91,7 @@ Array.union([], 1, null, [1, 2].values()) // [1, 2];
 ```
 
 * Added new prelude types: `AnyFunction`, `AnyOneArgFunction`, `Maybe`, `Either`
+* Added `ts-toolbelt`
 
 #### :house: Internal